backend/cloudFunctions/call_processor/requirements.txt
functions-framework==3.*
firebase-admin==6.*
google-cloud-firestore==2.*
google-cloud-storage==2.*
google-cloud-secret-manager==2.*
openai==1.*
Flask==2.*
google-auth-oauthlib==1.*
google-auth-httplib2==0.1.*
google-api-python-client==2.*
requests==2.*